+++ quizwhiz/js/main.js
@@ -4,8 +4,8 @@
 class QuizWhizApp {
     constructor() {
         // App version information
-        this.version = "3.8.2";
-        this.lastUpdated = "2025-08-09";
+        this.version = "3.9";
+        this.lastUpdated = "2025-08-10";
         
         // Expose version globally for footer
         window.APP_VERSION = this.version;